CMakeLists.txt: Add windows support for STM32_Programmer_CLI
I tested it and it actually builds. We should probably add a .gitignore
file so as not to destroy each other's toolchain configuration every
time we pull something from the remote repo. For the time being GDB is
acting funny if I don't clear the cache, so I should probably look more
into that.