- Remove cross-platform-software library.
The cross-platform-software library should not be included in comms software, as
most of the code was aimed towards atsam mcus, with stm32 specific code being added
using #ifdef STM32 /* ... */ #endif preprocessor commands, in specific comms compatible
branches. Now all stm32 specific code is implemented in the root project. The ecss-services
library is included as a submodule. Previously, it was a library of cross-platform-software, and
was checked out to a specific comms compatible commit in order to function properly. Now it is checked
out at master.

- Remove physical-layer library

Since a c++ implementation of error encoding and modulation is not required on the microcontroller (fpga task),
the library was removed.